+++ DD/src/components/Articles.tsx
@@ -1,8 +1,23 @@
+import { useState } from "react";
 import Article from "./Article";
 import './Articles.modules.css';
+import WoButton from "./Wobutton";
 
 export default function Articles() {
+    const [clicks, setClicks] = useState(0);
+    const [buttonText, setButtonText] = useState("View all");
+  
+    const handleClick = () => {
+      setClicks(clicks + 1);
+      if (clicks % 2 === 0) {
+        setButtonText("Hide");
+      } else {
+        setButtonText("View all");
+      }
+    };
+
     return (
+        <div style={{ display:  'flex', flexDirection: 'column', gap: '70px', overflow: 'hidden'}}>
         <div className="article-container">
             <Article 
             ImageSrc={"src/assets/images/Article1.png"} 
@@ -46,6 +61,37 @@ export default function Articles() {
                 </div>
             }
             />
+            {clicks % 2 === 1 && <><Article
+                ImageSrc={"src/assets/images/Article3.png"}
+                HeaderText={"Natural care for healthy facial skin"}
+                SubText={"A healthy lifestyle should start from now and also for your skin health. There are some..."}
+                ButtonText={<div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', gap: '15%' }}>
+                    Read more
+                    <img
+                        src="src/assets/images/Arrowright.svg"
+                        style={{ width: '14px', height: '9px' }} />
+                </div>} /><Article
+                    ImageSrc={"src/assets/images/Article3.png"}
+                    HeaderText={"Natural care for healthy facial skin"}
+                    SubText={"A healthy lifestyle should start from now and also for your skin health. There are some..."}
+                    ButtonText={<div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', gap: '15%' }}>
+                        Read more
+                        <img
+                            src="src/assets/images/Arrowright.svg"
+                            style={{ width: '14px', height: '9px' }} />
+                    </div>} /><Article
+                    ImageSrc={"src/assets/images/Article3.png"}
+                    HeaderText={"Natural care for healthy facial skin"}
+                    SubText={"A healthy lifestyle should start from now and also for your skin health. There are some..."}
+                    ButtonText={<div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', gap: '15%' }}>
+                        Read more
+                        <img
+                            src="src/assets/images/Arrowright.svg"
+                            style={{ width: '14px', height: '9px' }} />
+                    </div>} /></>
+            }
         </div>
+            <WoButton Text={buttonText} onClick={handleClick} />
+            </div>
     );
 }
